Initial commit of spreadsheet interaction.
Only committing this incomplete work because it has become clear that
developing on my workstation is the wrong thing, and developing on a GCE
instance is the right thing.  So I am saving my work here, pushing to my reop,
and then synching a GCE instance up with that same repo.